Make typedoc output deterministic and pretty.
Currently the issue with typedoc is every time it's run, it writes the current git blob hash to each file. Which means if I run it on every commit it will regenerate every single file, totally ruining the diff. Here I'm using `--disableSources` to stop this. Also, the html generates is on one single line. So again, if anything changes, it's very hard to see what's changed because the entire line changes. I'm using prettier to format the HTML files so that if I change the jsdoc for a function, only that line is changed in the typedoc output.