This command adds support for `rest` to be any type, not just string. It also adds a new `StringParser` type to simplify the types we show to the user.
As I try to make this library support input beyond just strings, the types are getting pretty complex. On success or failure, we return the rest of the input that hasn't been consumed. If the user's input was not a string, `rest` will likely not be a string also. So let's add support for that.

Since we assume the input parameter is a string unless otherwise specified, input is an optional type. Therefore, it's the last type in the list of genetics. This makes the types a little confusing to read, as you'd expect input to be the first type. Specifically, the types generated for the `or` combinator are tougher to read. I've added a new `StringParser` type to make the type easier to read. If this is a good idea is still to be determined.